Reduce the etcd surface area in the integration test to minimize the client dependency.
