+++ test/integration/BUILD
@@ -22,9 +22,5 @@ go_library(
         "//pkg/client/clientset_generated/release_1_5:go_default_library",
         "//pkg/client/clientset_generated/release_1_5/typed/core/v1:go_default_library",
         "//pkg/util/wait:go_default_library",
-        "//test/integration/framework:go_default_library",
-        "//vendor:github.com/coreos/etcd/client",
-        "//vendor:github.com/golang/glog",
-        "//vendor:golang.org/x/net/context",
     ],
 )
